refactor: 뒤로가기, 권한 컴포넌트 분리
